Add resource tracking to null.qubit (#7226)
**Context:**
For benchmarking, we need a method of precise resource tracking. This
method has to be compatible with code `@qjit` compiled using Catalyst
and should be able to track precise resources after various
transformation passes have already occurred.

See also: https://github.com/PennyLaneAI/pennylane-benchmarks/pull/76

**Description of the Change:**
Adds a new attribute to `null.qubit`, as well as an optional argument to
its `__init__` function. When this option is toggled on, `null.qubit`
will track precise gate counts when "executing" a circuit.

**Benefits:**
It is now possible to track precise gate counts (with or without gate
decomposition or other transformation passes) for measuring circuit
complexity.

**Possible Drawbacks:**
To be mutually compatible with the catalyst implementation, this
currently just saves the resources JSON to a file as opposed to neatly
returning an object. In the future, a better solution may be desired.
This implementation causes `null.qubit` to have a different capability
that other devices do not have. If this feature is useful elsewhere, it
might be wise to move resource tracking to the base Device class.

**Related GitHub Issues:**
[sc-88213]

---------

Co-authored-by: Christina Lee <christina@xanadu.ai>