Changing Sphinx minimum version (#7212)
**Context:**
Docs.yml uses python3.9 and sphinx version 3.5. It is currently a
blocker to dropping python 3.10 in pennylane. Upgrading sphinx to
version 8.1 allows for higher versions that are compatible with >Python
3.11 to be used later.

**Description of the Change:**
Sphinx is upgraded to version 8.1 and uses Python 3.10. References to
`<demos>` are updated to remove the `:doc:` prefix which is incompatible
with sphinx 8.1 and was broken previously.

**Benefits:**
Python 3.9 will be dropped.

**Possible Drawbacks:**

**Related GitHub Issues:**

---------

Co-authored-by: Rashid N H M <95639609+rashidnhm@users.noreply.github.com>