Manual cherry picking sync from latest rc to master (#7381)
This PR applies some changes to master coming from [this
PR](https://github.com/PennyLaneAI/pennylane/pull/7369/), which has been
merged in the release candidate branch for the bugfix release of
PennyLane 0.41.1.

More details can be found in the original PR description. This PR aims
to update the master branch with the changes currently present in the
stable version of PennyLane (0.41.1, no longer 0.41.0)