[Python Compiler] Creating the quantum dialect  (#7357)
**Context:** 

This is one of the first PRs implementing the new quantum compiler in
PennyLane.

This PR was originally opened in Catalyst:
https://github.com/PennyLaneAI/catalyst/pull/1691

**Description of the Change:** 

The changes introduced in this PR can be divided at a high level into 3
different steps:

*Step 1* 

Generated the Python dialect using `xdsl_tblgen.py`, using the JSON file
generated by `llvm-tblgen` starting from [this Catalyst
file](https://github.com/PennyLaneAI/catalyst/blob/main/mlir/include/Quantum/IR/QuantumOps.td).

*Step 2*

At the current stage, I commented the `assembly_format` attributes,
which generated an error, and added the `AttrSizedOperandSegments` and
`AttrSizedResultSegments` metadata

In fact, in the generated dialect, there is a class of errors that would
be good to fix in Catalyst. Erick is writing them as issues in Catalyst.
[Here](https://github.com/PennyLaneAI/catalyst/issues/1692) and
[here](https://github.com/PennyLaneAI/catalyst/issues/1690). Once these
issues are solved, we believe the whole quantum dialect would be easy to
generate and work relatively well to proceed to either other dialects or
the cancel-inverses transform.

*Step 3*

Writing some minimal tests for the dialect. These will be improved and
expanded (we will have multiple chats about this point).

**Benefits:**

**Possible Drawbacks:**

**Related GitHub Issues:** None.

**Related Shortcut Story:** [sc-89798]

---------

Co-authored-by: Isaac De Vlugt <34751083+isaacdevlugt@users.noreply.github.com>