Update release to adhere to PyPI BDF requirements (#7382)
**Context:** The PyPI release process requires more strict adherence to
the
https://packaging.python.org/en/latest/specifications/binary-distribution-format/
guidelines. `PennyLane` the package name is not compliant with this
naming, and requires the released wheels to be named `pennylane`. Older
versions of setuptools were less strict about this naming, and forwarded
what was requested in the package metadata. Upgrade setuptools and wheel
build infrastructure ensures that we adhere with the wheel naming
expectations of PyPI.
Additionally, this PR also removes the deprecated license classifier,
removing a warning of `SetuptoolsDeprecationWarning: License classifiers
are deprecated`.

**Description of the Change:** Replaces the direct call to setup.py for
wheel building with a call to https://github.com/pypa/build

**Benefits:** Ensures adherence to Python packaging expectations coming
into enforcement from PyPI. Fixes future failure expected during
release.

**Possible Drawbacks:**

**Related GitHub Issues:**