Allow returning diagonalized parametric MCMs (#7387)
**Context:**

When we diagonalize the parametric mid-circuit measurements to get
analytic results to validate our protocols, we swap them out with new,
updated measurements. This is currently updated in the conditionals, but
not on the measurements. That means that returning parametric mcms when
using the `diagonalize_mcms` transform currently fails.

**Description of the Change:**

We also update the measurements in the transform - if any
MeasurementProcess contains MCMs, we update them based on the existing
dictionary that's tracking updated measurement values, and create a copy
of the MP with the new MeasurementValue(s) instead of the original ones.

**Benefits:**

We can do this:

```
@diagonalize_mcms
@qml.qnode(dev)
def circ():
    qml.H(0)
    m = measure_x(0, reset=True)
    return qml.expval(m)
```

[sc-90609]