Reduce memory and compute usage in uniformly controlled rotations (`MottonenStatePreparation`) (#7377)
**Context:**
In `MottonenStatePreparation`, there is a step where the desired phases
need to be transformed into rotation angles for the Gray code-based
rotation sequence that implements the phases.
The sequence is shown in Fig. 2, the angle transformation in Eq. (3) of
[Möttönen et al.](https://arxiv.org/pdf/quant-ph/0407010).

Currently, the transformation is done by explicitly constructing the
matrix $M$ from Eq.(3) and multiplying it to the angles.
This is not only painstakingly slow but also requires quadratic memory
size compared to the input and output. If we care about preparing larger
state vectors, this limits the computation to about 15 qubits, which it
does not have to do.

Instead, one can compute the angles with a fast Walsh-Hadamard transform
(FWHT) and subsequent permutations. These permutations act as if we
would apply a ladder of CNOT gates to the angles, when interpreting them
as a quantum state.

**Description of the Change:**
Change `compute_theta`, which implements the transform, from the
matrix-based method to the FWHT-based one.

**Benefits:**
Roughly reduces memory requirement from `O(d^2)` to `O(d)`, where `d` is
the input (and output) dimension.
Reduces compute time _heavily_:

![image](https://github.com/user-attachments/assets/de083bcb-d5ff-4406-8dca-07395c13396d)


**Possible Drawbacks:**
N/A

**Related GitHub Issues:**
#7378 
[sc-90387]

---------

Co-authored-by: Korbinian Kottmann <43949391+Qottmann@users.noreply.github.com>
Co-authored-by: Isaac De Vlugt <34751083+isaacdevlugt@users.noreply.github.com>