fix(openfermion): Use identity qubit map when none is provided (#7332)
**Context:**

The documentation of the `to_opernfermion` function
(https://docs.pennylane.ai/en/stable/_modules/pennylane/qchem/convert_openfermion.html#to_openfermion)
states that if no qubit-to-wire dictionary is provided then the identity
map is used. This does not seem to be the case, however, when the
operators only act on a subset of the qubits. For instance the snippet:

```python
sum_op = 1.2 * qml.X(1) + 2.4 * qml.Z(2) 
of_qubit_op = qml.to_openfermion(sum_op) 

# returns 
# (1.2+0j) [X0] +
# (2.4+0j) [Z1]

# expected
# (1.2+0j) [X1] +
# (2.4+0j) [Z2]
```

If this is expected behavior, I think the documentation should be
updated.

**Description of the Change:**

This PR implements the identity map when no mapping dictionnary is
provided.

**Benefits:**

**Possible Drawbacks:**

**Related GitHub Issues:**

---------

Co-authored-by: Diksha Dhawan <40900030+ddhawan11@users.noreply.github.com>
Co-authored-by: Isaac De Vlugt <34751083+isaacdevlugt@users.noreply.github.com>
Co-authored-by: soranjh <soran.jahangiri@gmail.com>