Upgrading ReadTheDocs and RTD runners to ubuntu24 (#7396)
**Description of the Change:**
Upgrading the readthedocs.yml and rtd.yml action runner to Ubuntu-24.04
from Ubuntu22.04; 22.04 is in end of life. The docs.yml runner will also
be pinned to ubuntu-24.04 to match the other docs checks.

**Benefits:**
The readthedocs, rtd, and docs runners gets updated.

**Possible Drawbacks:**
There could be potential compatibility issues with sphinx and other
packages. None were found in the PR run.

**Related GitHub Issues:**

---------

Co-authored-by: Pietropaolo Frisoni <pietropaolo.frisoni@xanadu.ai>