Add transform to change measurements to samples/counts (#7317)
**Context:**

Many backends return measurements only as raw samples in the
computational basis (stored either as a single-shot array like `sample`,
or a total number like `counts`).

For these backends, a tape that requests other types of measurements
(`expval`, `probs`, etc) needs to be converted to a tape that requests
only `sample`/`counts`, and then post-processed to return the expected
output to the user.

This functionality already exists in Catalyst, but hasn't been added to
PennyLane.

**Description of the Change:**
We add the corresponding function to PennyLane, in the form of
`measurements_from_counts` and `measurements_from_samples` transforms.

**Benefits:**
We provide a generic transform for this functionality (currently
implemented on a case-by-case basis in plugins) for future plugin
development and for users.

This will be useful in the MBQC circuit conversion pipeline.


[sc-89543]

---------

Co-authored-by: Isaac De Vlugt <34751083+isaacdevlugt@users.noreply.github.com>
Co-authored-by: Christina Lee <christina@xanadu.ai>
Co-authored-by: Yushao Chen (Jerry) <chenys13@outlook.com>