Pinning `cotengra` (#7423)
**Context:** `cotengra` released recently a [newer
version](https://github.com/jcmgray/cotengra/releases/tag/v0.7.3) and
this broke the CI

**Description of the Change:** We temporarily pin the installed
`cotengra` version. Still, we should update the `DefaultTensor` device
to be compatible with the latest `cotengra`, because now we are manually
downgrading it after the installation of `quimb`, relying on pip’s
resolution order, which can be fragile.

**Benefits:** CI passes.

**Possible Drawbacks:** Mismatch between the `cotengra` version that
`quimb` expects and the one manually pinned here.

**Related GitHub Issues:** None.

[sc-91237]

---------

Co-authored-by: JerryChen97 <chenys13@outlook.com>