Modify resource tracking to write to distinct filenames (#7392)
**Context:**
The new resource tracking feature introduced in #7226 logs resource
tracking information to a json file. This can cause issues when the file
gets clobbered after multiple runs.

**Description of the Change:**
This PR changes the behaviour of resource tracking to print resource
information to a json which has a timestamp as a suffix instead of
writing to a static filename. This behaviour is consistent with
Catalyst.

**Benefits:**
Avoids potential issues with file clobbering.

**Possible Drawbacks:**
Parsing can no longer be done for a static filename, the written file
must be searched for.

**Related GitHub Issues:**
#7226
[sc-88213]