Add `qml.to_openqasm` transform (#7393)
This PR adds a transform to make the conversion of a quantum circuit
into the corresponding OpenQASM 2.0 program easier and more
user-friendly. The `qml.to_openqasm` function is implemented in a new
`openqasm` module under the `qml.transforms` subpackage. Not sure this
is the right location in the Pennylane repository.

A quite extensive set of unit and integration tests for the
`QuantumScript.to_openqasm` method is already available in
[test_qasm.py](https://github.com/PennyLaneAI/pennylane/blob/master/tests/circuit_graph/test_qasm.py).
I will add a detailed description of what the transform is doing in the
docstring, including a simple usage example.

[sc-83360]

---------

Co-authored-by: Yushao Chen (Jerry) <chenys13@outlook.com>
Co-authored-by: Pietropaolo Frisoni <pietropaolo.frisoni@xanadu.ai>
Co-authored-by: Astral Cai <astral.cai@xanadu.ai>
Co-authored-by: Isaac De Vlugt <34751083+isaacdevlugt@users.noreply.github.com>