enforce subset of `templates` submodules as "auxiliary" modules (#7437)
**Context:**

Following our PL module structure, the goal of this PR is to isolate
straight forward submodules of `templates` as a `auxiliary` modules.

**Description of the Change:**

⚠️ This is a rather large and complex module which is unfortunately
interconnected to the `ops` module. Due to this, there are circular
import errors when trying to clean-up the `subroutines` and
`state_preparations` submodules.

This PR does the following,
- Updates `tach.toml` to break up `templates` into individual submodules
(allows for submodule level enforcement)
- Address any errors that arise

**Benefits:** Module dependency enforcement.

**Possible Drawbacks:** None identified.

[sc-91368]