[Decomposition] Clean up custom logic for adjoint and pow (#7352)
**Context:**

This is a follow-up PR to
https://github.com/PennyLaneAI/pennylane/pull/7347

**Description of the Change:**

There was previously a lot of custom logic for retrieving decomposition
rules of symbolic operators. With the infrastructure added in
https://github.com/PennyLaneAI/pennylane/pull/7347, this PR is able to
clean up most of it.

- We no longer hard-code a list of operators whose adjoint is an
instance of the same type. Instead, each operator with a custom adjoint
decomposition registers a decomposition rule for the adjoint of itself.
For example, `Hadamard` was previously hardcoded as an operator that is
its own adjoint, now there is a decomposition rule for
`"Adjoint(Hadamard)"` which produces an instance of the base operator.
- We removed the constraint that powers of an operator must be integers.
Non-integer powers can be supported by adding custom decomposition rules
for `"Pow(SomeOp)"`. All operators in PL which overrides `pow` now has a
custom decomposition rule registered under the power of itself.
- Defined some general stock decomposition rules that can be registered
under any operator. For example, a `self_adjoint` decomposition rule is
defined, and it is registered with all operators that are self-adjoint
(to avoid having duplicate code)

**Related Shortcut Stories:**

[sc-90114]

---------

Co-authored-by: Pietropaolo Frisoni <pietropaolo.frisoni@xanadu.ai>