Support for weighted gates in graph decomposition algorithm (#7389)
------------------------------------------------------------------------------------------------------------

**Context:** In short, a decomposition system based on a graph traversal
using Dijkstra's algorithm is applied to find the most optimal gate
decompositions for non-native gates. The nodes in the graph represent
either decompositions, operations in the target gateset, or higher order
operations. The edges show either membership of a native op in a
decomposition, or equivalence of a decomposition to a higher order
operation. The cost of an edge between a native operation and a
decomposition is the number of additional gates found in the
decomposition. The cost of an edge between a decomposition and an
equivalent higher order op is zero. For more context, see
[[sc-84664](https://app.shortcut.com/xanaduai/story/84664)].

**Description of the Change:** Work in progress on implementing support
for relatively weighted gates within a target gateset. The weights are
provided by the user, and influence what gate decompositions are chosen
by the graph decomposition algorithm. The implementation is such that
the weights of the gates are represented by weights on the edges leading
to and from native gates.

**Benefits:** Allows a user to more closely model the behaviour of their
backends, and capture the relative cost of executing certain gates such
as non-Clifford gates which are native, but more expensive.

**Possible Drawbacks:** The representation of weights in this way is
possibly not intuitive.

**Related GitHub Issues:**

---------

Co-authored-by: Astral Cai <astral.cai@xanadu.ai>
Co-authored-by: Pietropaolo Frisoni <pietropaolo.frisoni@xanadu.ai>
Co-authored-by: Yushao Chen (Jerry) <chenys13@outlook.com>
Co-authored-by: Isaac De Vlugt <34751083+isaacdevlugt@users.noreply.github.com>