`qml.set_shots` is available (#7337)
**Context:**
Overall, it's not a good idea to initialize a device like now in
PennyLane that is bounded with `shots=`. It would be a better
abstraction if we can separate these two concepts. Therefore this PR
aims to bring a new transformer to implement such, but there must be
lots of conflicts inside our current pipeline, which requires extra
attention in another PR. The biggest issue will be that the allowed diff
methods between analytic and finite-shots are not compatible with each
other.

**Description of the Change:**
A new transformer `set_shots` shall enable users to do the following:
```python
device = qml.device("default.qubit", wires=2)

@partial(set_shots, shots=1000)
@qml.qnode(device)
def circuit(x):
    qml.RX(x, wires=0)
    qml.RY(x, wires=1)
    qml.CNOT(wires=[0, 1])
    return qml.expval(qml.PauliZ(0)), qml.expval(qml.PauliZ(1))
```

**Benefits:**

**Possible Drawbacks:**
The actual runnability is dependent on
https://github.com/PennyLaneAI/pennylane/pull/7358

**Related GitHub Issues:**
[sc-89988]

---------

Co-authored-by: Christina Lee <christina@xanadu.ai>
Co-authored-by: Isaac De Vlugt <34751083+isaacdevlugt@users.noreply.github.com>
Co-authored-by: Josh Izaac <josh146@gmail.com>