use `exceptions` module in source code (#7292)
**Note to Reviewers:** 

1. Code coverage is complaining about a line that is currently uncovered
on master - can be ignored for the scope of this PR.
2. VSCode's refactor feature was used to correct the imports for all of
these exceptions.

---

**Context:**

https://github.com/PennyLaneAI/pennylane/pull/7205 added the new
`exceptions` module but doesn't use it internally.
This is because Catalyst and Lightning used the top-level imports
leading to failures in the external tests CI.

With all of the plug-ins / Catalyst / Lightning updated, we can now
update our source code.

**Description of the Change:**

- Update to use that new module rather than top level imports
(documentation as well)
- Top-level access will now raise a deprecation warning

**Benefits:** Code is organized.

**Possible Drawbacks:** Will have deprecation warnings now which could
mess up the PTM if something was missed.

[sc-88956]

---------

Co-authored-by: Mudit Pandey <mudit.pandey@xanadu.ai>