Transform to convert to MBQC formalism (#7355)
**Context:**

We have a transform to go from an arbitrary circuit to the gateset
compatible with textbook MBQC. Now we want to convert that the MBQC
formalism, expressed using resource states, parametrized MCMs and
byproduct corrections.

**Description of the Change:**

We add a transform that performs a simple conversion from that gate-set
to the MBQC formalism. "Simple" here means:

- No consolidation of Pauli operations and/or byproduct operations
- Byproducts are applied online after each measurement protocol is
complete, no offline corrections
- we use the qubit manager to allocate qubits on the tape before
execution (not capture compatible, we unroll all program structure)

This is a very direct, naive conversion to the MBQC formalism. But it
does the thing, so it's a good place to start - and we can use it as a
baseline for quantifying future improvements in our circuit conversion.

**Benefits:**
We can convert a circuit to a simple, un-optimized implementation of the
same circuit in the MBQC formalism.

[sc-88533]

---------

Co-authored-by: Isaac De Vlugt <34751083+isaacdevlugt@users.noreply.github.com>
Co-authored-by: Christina Lee <christina@xanadu.ai>
Co-authored-by: Yushao Chen (Jerry) <chenys13@outlook.com>
Co-authored-by: Joey Carter <joseph.carter@xanadu.ai>