[FTQC]Add Clifford commutation method to Pauli Tracking (#7444)
### Before submitting

Please complete the following checklist when submitting a PR:

- [x] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      test directory!

- [x] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [x] Ensure that the test suite passes, by running `make test`.

- [x] Add a new entry to the `doc/releases/changelog-dev.md` file,
summarizing the
      change, and including a link back to the PR.

- [x] The PennyLane source code conforms to
      [PEP8 standards](https://www.python.org/dev/peps/pep-0008/).
We check all of our code against [Pylint](https://www.pylint.org/).
      To lint modified files, simply `pip install pylint`, and then
      run `pylint pennylane/path/to/file.py`.

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:**

[sc-91490]

This PR implements the conjugate rules for a set of Clifford gates
(`qml.H`, `qml.S`, `qml.CNOT`) and Paulis based on the fact that a
Cliiford gate ($C$) can conjugate a Pauli to a new Pauli, i.e.,
$P_{new}C = CP_{old}$.

**Conjugate rules**
- $CNOT$ gate

||$X_c$|$Z_c$|$X_t$|$Z_t$|
|:--:|:--:|:--:|:--:|:--:|
|$\pm$|$+$|$+$|$+$|$+$|
|$control$|$X$|$Z$|$I$|$Z$|
|$target$|$X$|$I$|$X$|$Z$|
 
- $Hadamard$ gate
 
||$X$|$Z$|
|:--:|:--:|:--:|
|$\pm$|$+$|$+$|
|$target$|$Z$|$X$|
 
- $S$ gate
 
||$X$|$Z$|
|:--:|:--:|:--:|
|$\pm$|$i$|$+$|
|$target$|$XZ$|$Z$|


**Description of the Change:**

**Benefits:**

**Possible Drawbacks:**

**Related GitHub Issues:**

---------

Co-authored-by: Joey Carter <joseph.carter@xanadu.ai>
Co-authored-by: lillian542 <38584660+lillian542@users.noreply.github.com>