[Decomposition] Decomposition rules for ControlledQubitUnitary (#7371)
**Description of the Change:**

- Existing decomposition rules for controlled single-qubit unitaries
(`ctrl_decomp_bisect` and `ctrl_decomp_zyz`) are re-written as
capture-compatible quantum functions with registered resource
requirements, then added to the library of decomposition rules under
`ControlledQubitUnitary`.
- An additional `controlled_two_qubit_unitary_rule` is added which
applies control to the decomposition of the two-qubit unitary, also
added to the list of decomposition rules under `ControlledQubitUntiary`.
- Decomposition rules for controlled qubit unitaries are all moved from
`ops/op_math/controlled_decompositions.py` to
`ops/op_math/decompositions/controlled_decompositions.py`
- A `to_controlled_qubit_unitary` decomposition is added to the
decomposition graph that considers decomposing any controlled
single-qubit operator to a controlled qubit unitary. In the old system,
this was a last resort for when the base operator does not have a
decomposition (which is never the case in the new system), so we would
always consider this as an option.

**Related GitHub Issues:**

[sc-89080]

---------

Co-authored-by: Pietropaolo Frisoni <pietropaolo.frisoni@xanadu.ai>
Co-authored-by: Yushao Chen (Jerry) <chenys13@outlook.com>
Co-authored-by: Andrija Paurevic <46359773+andrijapau@users.noreply.github.com>