Mark flaky FTQC test as flaky (#7511)
**Context:**
We have a full, device execution E2E test for an execution pipeline that
converts a tape to the MBQC formalism and gets results, and it only has
500 shots, because it's slow. So it's flaky, and it's been causing test
failures. Seeding the device was not sufficient to fix it.

**Description of the Change:**
We mark the test as flaky, and add a comment explaining. 

**Benefits:**
We keep the full E2E test without blocking CI. Unit tests can validate
we are doing what we *think* we should be doing, for the steps we
identify as important to test, and they're great. But if we've made a
mistake in our understanding of the protocol's implementation, or if we
are missing a step that is relevant to test, the unit tests won't tell
us. This test is a sanity check. It isn't exhaustive either, but it's
valuable information when we are working with a lot of experimental
features and new procedures.

Eventually tests like this could be moved into a separate test suite
that runs less frequently as a sanity check, instead of being included
in all CI runs.

**Possible Drawbacks:**
We want to reduce the number of flaky tests in our repo, and generally
move away from full device execution for tests, so this isn't fully
in-line with our testing philosophy. However, it was discussed in the
FTQC team meeting and we got a go-ahead from @mlxd and @Alex-Preciado
for this case.