Add ROWCOL CNOT routing algorithm (#7394)
**Context:**
A pure CNOT circuit can be described entirely via its parity matrix. In
#7229, a function is added to `qml.labs` to compute the parity matrix
for a CNOT circuit. The answer there is unique: A CNOT circuit maps to
one unique parity matrix.
The reverse task of extracting, or synthesizing, a CNOT circuit that
implements a given parity matrix is accomplished by a series of
algorithms in the literature, a better-known one being ROWCOL introduced
[here](https://arxiv.org/abs/1910.14478) and described in the
compilation hub
[here](https://pennylane.ai/compilation/rowcol-algorithm).
Note that there are infinitely many CNOT circuits that produce the same
parity matrix, so that the answer to this task is never unique, and
different algorithms will return CNOT circuits with different gate
counts and depths.

**Description of the Change:**
Add a `rowcol` function to `qml.labs` that takes a parity matrix $P$ and
a qubit connectivity graph as inputs and returns a CNOT circuit that has
the parity matrix $P$ and that respects the connectivity.

**Benefits:** 
CNOT routing and circuit optimization capabilities.

**Possible Drawbacks:**
N/A

**Related GitHub Issues:**

---------

Co-authored-by: qottmann <korbinian.kottmann@gmail.com>
Co-authored-by: Korbinian Kottmann <43949391+Qottmann@users.noreply.github.com>
Co-authored-by: Guillermo Alonso-Linaje <65235481+KetpuntoG@users.noreply.github.com>