Fix `qml.TShift` power method (#7505)
This PR implements a fix for issue #6355 (`pow` method for non
parametric operators acting on a single qutrit).

The method `Operator.pow` is supposed to return a list with a simplified
"decomposition" of the given operator. For example, in the case of
`TShift` we have `TShift**3` = `Identity` and so `TShift.pow(3)` returns
an empty list. This PR handles the general case `TShift.pow(z)`
returning a list with `z%3` copies of the given operator (actually the
only relevant case is when `z=2`).

This fix should work also for the `qml.TClock` and `qml.TAdd` operators.

[sc-75304]

---------

Co-authored-by: Pietropaolo Frisoni <pietropaolo.frisoni@xanadu.ai>
Co-authored-by: Yushao Chen (Jerry) <chenys13@outlook.com>