Simple QASM 3.0 gates can be executed (no pre-compute) (#7495)
------------------------------------------------------------------------------------------------------------

**Context:** We would like to get a basic start on an interpreter for
OpenQASM 3.0 which return a callable that
can be queued into a QNode and then executed when the QNode is used.
This is an alternative version of the PR here:
[(#7432)](https://github.com/PennyLaneAI/pennylane/pull/7432) which has
no pre-compute step. [sc-90384]

**Description of the Change:** This change includes basic support for
simple usages of gates with directly provided parameters and tests which
verify this functionality. The new python file `qasm_interpreter.py`
includes a class which extends the QASMVisitor class provided by
openqasm3 in order to traverse a simple AST and build the callable from
it.

**Benefits:** This is the start of a project which will eventually allow
users to convert arbitrary OpenQASM 3.0 programs into Pennylane QNodes.

**Possible Drawbacks:** This approach is laser focused on interpreting
simple series of gates and does not include support for basic things
like using expressions or references instead of literals to parameterize
gates, etc. All of these necessary things will be added in future PRs
according to [sc-90379]

**Related ShortCut Stories:** [sc-90384]

---------

Co-authored-by: Astral Cai <astral.cai@xanadu.ai>
Co-authored-by: Christina Lee <christina@xanadu.ai>
Co-authored-by: Isaac De Vlugt <34751083+isaacdevlugt@users.noreply.github.com>