Fixing qml links (#7453)
**Context:**
Many of the links to classes listed in the [qml
page](https://docs.pennylane.ai/en/latest/code/qml.html#classes) of the
documentation website on latest aren't working. The class docs are still
there if you type in the URL manually. The links were broken after the
docs.yml sphinx upgrade from Python3.9 to Python3.10 and Sphinx 3.5 to
8.1.

The upgrade created warnings like below so `:no-index:` was used in the
`class.rst` template to remove the index of duplicate instances.

```
/root/pennylane/pennylane/bose/bosonic.py:docstring of pennylane.bose.bosonic.BoseSentence:1: WARNING: duplicate object description of pennylane.bose.bosonic.BoseSentence, other instance in code/api/pennylane.BoseSentence, use :no-index: for one of them

/root/pennylane/pennylane/bose/bosonic.py:docstring of pennylane.bose.bosonic.BoseWord:1: WARNING: duplicate object description of pennylane.bose.bosonic.BoseWord, other instance in code/api/pennylane.BoseWord, use :no-index: for one of them
```

**Description of the Change:**
Removed duplicate occurrences of the classes:

- pennylane.bose.bosonic.BoseSentence
- pennylane.bose.bosonic.BoseWord
- pennylane.devices.tracker.Tracker
- pennylane.fermi.fermionic.FermiA
- pennylane.fermi.fermionic.FermiC
- pennylane.fermi.fermionic.FermiSentence
- pennylane.fermi.fermionic.FermiWord
- pennylane.io.qualtran_io.FromBloq

**Benefits:**
Fix broken qml links

**Possible Drawbacks:**
None

**Related GitHub Issues:**
[PR-7348](https://github.com/PennyLaneAI/pennylane/pull/7348)

---------

Co-authored-by: justinpickering <justin.r.pickering@gmail.com>