Fix `map_to_standard_wires` to work with work wires (#7581)
**Context:**
`qml.tape.QuantumScript.map_to_standard_wires` maps the circuit wires to
some standard choice of wires, with the following conventions:
- The new wires are consecutive integers starting at 0.
- Wires used in operations come first, including those that _also_ are
acted on by measurements.
- Wires _exclusively_ acted on by measurements come after operations'
wires.

If the wire order of the circuit already is compatible with these
conventions, no mapping is performed.

Work wires, however, are not considered at all in this method.
As work wires of some important operations (like `MultiControlledX`) are
not part of the operation's wires, they are disregarded entirely by
`map_to_standard_wires`.

**Description of the Change:**
Take `work_wires` into account, extending the above conventions by the
following point:
- wires _exclusively_ used as work wires come after exclusive
measurement wires.

Also, this PR adds test for the method, which we did not have as unit
tests.

**Benefits:**
- Fixes #7580 (note that this includes circuit execution as soon as an
exclusive work wire is flying around in the circuit).
- If no decomposition involving the work wires is performed, the
exclusive work wires are trailing wires that can be "ignored" away
easily, i.e. there are no mysterious gaps in the standard wire ordering
just because there _would_ be work wires _if_ we did the decomposition.

**Possible Drawbacks:**
I first thought that this introduces some non-canonical choices of how
to handle work wires as part of operators or not.
But this is actually not the case, because `map_to_standard_wires`
basically just creates a `wire_map` that is used for mapping in
`qml.map_wires`. So the choices for how to treat work wires, for example
when an operation is not actually decomposed, are not affected, because
they are made by `map_wires`.
This PR only changes _where_ the `work_wires` go if they are mapped by
some other method (like `qml.map_wires`).

**Related GitHub Issues:**
#7580
[sc-92378]

---------

Co-authored-by: Pietropaolo Frisoni <pietropaolo.frisoni@xanadu.ai>