Metric tensor compatibility with catalyst (#7528)
**Context:**

We are wanting to make our optimizers compatible with catalyst.
`QNGOptimizer` requires the metric tensor, so we first need to make the
metric tensor compatible with catalyst.

**Description of the Change:**

Adds `indices_are_sorted` and `unique_indices` keyword arguments to
`scatter_element_add`. Loop over indices independently when they are not
sorted.

**Benefits:**

`metric_tensor` can be computed with catalyst.

**Possible Drawbacks:**

**Related GitHub Issues:**

---------

Co-authored-by: Andrija Paurevic <46359773+andrijapau@users.noreply.github.com>
Co-authored-by: Yushao Chen (Jerry) <chenys13@outlook.com>
Co-authored-by: Simone Gasperini <simone.gasperini4@unibo.it>